+++ lib/html2pdf/html2pdf.rb
@@ -1,6 +1,7 @@
 require "open3"
 require "fileutils"
 require_relative "../html2pdf"
+require_relative "./configuration"
 module Html2Pdf
   class << self
     # Batch convert to pdf using `wkhtmltopdf` tool
@@ -19,23 +20,15 @@ module Html2Pdf
     # @param filename input filename
     def to_pdf(filename)
       fail "Invalid input file #{filename}" unless File.exist?(filename)
-      # TODO: allow custom configuration
+      wkhtmltopdf   = Html2Pdf.configuration.options[:wkhtmltopdf]
+      page_settings = Html2Pdf.configuration.options[:page_settings]
       command = [
-        "wkhtmltopdf",
-        "--margin-top 4",
-        "--margin-bottom 4",
-        "--margin-left 4",
-        "--margin-right 4",
-        '--header-center "[webpage] :: [page]/[topage]"',
-        "--header-spacing 1",
-        "--header-font-size 8",
-        "--header-line",
-        "--footer-spacing 1",
-        "--footer-font-size 8",
-        "--footer-line",
-        "#{filename}",
+        wkhtmltopdf,
+        *page_settings,
+        filename,
         "#{filename}.pdf",
-        "> /dev/null"]
+        "> /dev/null"
+      ]
       _stdin, _stderr, status = Open3.capture3(command.join(" "))
       fail "Problem processing #{filename}" unless status.success?
     end